Move config validation to orchestrator
